Fix REST API spec and database lat lng data types
- Changes the data types for latitude and longitude to int in the spec
and database to ensure that no floating point inconsistencies can cause
issues across systems
- Fixes some other openapi spec formatting issues to match the result
of importing into the swagger editor

+++ subscriber/simple_supply_subscriber/database.py
@@ -53,8 +53,8 @@ CREATE_RECORD_LOCATION_STMTS = """
 CREATE TABLE IF NOT EXISTS record_locations (
     id               bigserial PRIMARY KEY,
     record_id        varchar,
-    latitude         numeric(9,6),
-    longitude        numeric(9,6),
+    latitude         bigint,
+    longitude        bigint,
     timestamp        bigint,
     start_block_num  bigint,
     end_block_num    bigint
